refactor(HighlightComponent): remove HighlightComponent usage
The HighlightComponent is no longer needed on the Home page due to changes in the
text search functionality. The component has been removed to prevent redundancy and
maintain a clean and efficient codebase.

refactor(TextSearch): use styled-components for highlighting

Refactor the TextSearch component to utilize styled-components for better styling
management. This change allows for more maintainable and scalable CSS practices,
enhancing the component's visual customization capabilities.